// backend/src/types/application.ts
export interface Application {
    application_id: number;
    name: string;
    description?: string;
    company_name?: string;
    email?: string;
    telefon?: string;
    pdf_src?: string;
    expire_date?: string;
    creation_date: string;
    lastupdate_date: string;
    applicationtype: string;
}